+++ packages/component-library/components/Dropdown/index.ts
@@ -1,1 +1,2 @@
 export { default as Dropdown } from "./Dropdown"
+export { DropdownProps } from "./Dropdown"