+++ draft-packages/button/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@1.5.5...@kaizen/draft-button@1.5.6) (2020-06-17)
+
+**Note:** Version bump only for package @kaizen/draft-button
+
+
+
+
+
+## [1.5.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@1.5.4...@kaizen/draft-button@1.5.5) (2020-06-16)
+
+**Note:** Version bump only for package @kaizen/draft-button
+
+
+
+
+
+## [1.5.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@1.5.3...@kaizen/draft-button@1.5.4) (2020-06-12)
+
+**Note:** Version bump only for package @kaizen/draft-button
+
+
+
+
+
 ## [1.5.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@1.5.2...@kaizen/draft-button@1.5.3) (2020-06-12)
 
 **Note:** Version bump only for package @kaizen/draft-button