+++ draft-packages/collapsible/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-collapsible@1.3.7...@kaizen/draft-collapsible@1.3.8) (2020-06-17)
+
+**Note:** Version bump only for package @kaizen/draft-collapsible
+
+
+
+
+
+## [1.3.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-collapsible@1.3.6...@kaizen/draft-collapsible@1.3.7) (2020-06-16)
+
+**Note:** Version bump only for package @kaizen/draft-collapsible
+
+
+
+
+
+## [1.3.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-collapsible@1.3.5...@kaizen/draft-collapsible@1.3.6) (2020-06-12)
+
+**Note:** Version bump only for package @kaizen/draft-collapsible
+
+
+
+
+
 ## [1.3.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-collapsible@1.3.4...@kaizen/draft-collapsible@1.3.5) (2020-06-12)
 
 **Note:** Version bump only for package @kaizen/draft-collapsible