+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.2.7...@kaizen/draft-guidance-block@2.2.8) (2020-06-17)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.2.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.2.6...@kaizen/draft-guidance-block@2.2.7) (2020-06-16)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.2.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.2.5...@kaizen/draft-guidance-block@2.2.6) (2020-06-12)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
 ## [2.2.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.2.4...@kaizen/draft-guidance-block@2.2.5) (2020-06-12)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block