+++ draft-packages/table/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.11...@kaizen/draft-table@1.6.12) (2020-06-18)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [1.6.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.10...@kaizen/draft-table@1.6.11) (2020-06-18)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [1.6.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.9...@kaizen/draft-table@1.6.10) (2020-06-18)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [1.6.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.8...@kaizen/draft-table@1.6.9) (2020-06-17)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [1.6.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.7...@kaizen/draft-table@1.6.8) (2020-06-16)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [1.6.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.6...@kaizen/draft-table@1.6.7) (2020-06-12)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
 ## [1.6.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@1.6.5...@kaizen/draft-table@1.6.6) (2020-06-12)
 
 **Note:** Version bump only for package @kaizen/draft-table