+++ draft-packages/zen-off-canvas/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.6.19...@kaizen/draft-zen-off-canvas@1.6.20) (2020-06-17)
+
+**Note:** Version bump only for package @kaizen/draft-zen-off-canvas
+
+
+
+
+
+## [1.6.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.6.18...@kaizen/draft-zen-off-canvas@1.6.19) (2020-06-16)
+
+**Note:** Version bump only for package @kaizen/draft-zen-off-canvas
+
+
+
+
+
+## [1.6.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.6.17...@kaizen/draft-zen-off-canvas@1.6.18) (2020-06-12)
+
+**Note:** Version bump only for package @kaizen/draft-zen-off-canvas
+
+
+
+
+
 ## [1.6.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.6.16...@kaizen/draft-zen-off-canvas@1.6.17) (2020-06-12)
 
 **Note:** Version bump only for package @kaizen/draft-zen-off-canvas